[API-645] Add asset attachment & logo management for v2 (#84)
* [API-645] Document attachment endpoints for assets (v2)

* [API-645] Document logo endpoint for assets (v2)

* [API-645] Use single quotes for cURL CLI arguments

* [API-645] Capitalize headers, Location header

* [API-645] Add note on how to obtain attachment identifiers

* [API-645] Update description of `attachments` form field

It should be the content of the file to upload as an attachment, not the
path to the file, as seen in the cURL example - the notation in the cURL
example takes care of reading the file for you.

* [API-645] Clarify how the `Location` header can be used

Namely, only to DELETE the attachment. Getting an attachment directly is
(not yet?) supported.

* [API-645] Update line formatting for asset attachments

* [API-645] Remove file type requirement for upload attachment

The requirements are quite lax, various types of files are allowed. For
documentation simplicity we will not provide an exhaustive list, most
file types a non-malicous user would want to upload are supported.

* [API-645] Clarify uploading a logo replaces the old one

* [API-645] Update the v2 changelog

Co-authored-by: M. Serhat Dundar <msdundars@gmail.com>