[API-1543] Pin GitHub Actions to SHAs (#136)
Pin all GitHub Actions we use to the commit SHA of the latest
release in the same major version as we were using. Note, for
the `vale-action`, the use of `reviewdog` referenced a branch and v2.0.1
points to an older commit on that branch.

This should improve reproducability and security.